move snap creation to the snapping class
