+++ core/lib/snapping.js
@@ -7,6 +7,35 @@ export class Snapping {
     this.active = false;
   }
 
+  /**
+   * Get snap point and draw to the scene
+   * @param {scene} scene object //TODO: passing scene is hacky. Find a cleaner way
+   */
+  snap(scene) {
+    const snapPoint = this.getSnapPoint(scene);
+    if (snapPoint) {
+      this.addSnapPoint(snapPoint, scene);
+    }
+  }
+
+  addSnapPoint(snapPoint, scene) {
+    // Draw a circle to highlight the snap.
+    const CentrePoint = new Point(snapPoint.x, snapPoint.y);
+    const radiusPoint = new Point(snapPoint.x, snapPoint.y + (5 / scene.core.canvas.getScale()));
+    const snapCirclePoints = [CentrePoint, radiusPoint];
+
+    const data = {
+      points: snapCirclePoints,
+      colour: scene.core.settings.snapcolour.toString(),
+    };
+
+    const item = scene.core.commandManager.createNew('Circle', data);
+    scene.tempItems.push(item);
+
+    // Move the mouse to the closest snap point so if the mouse if clicked the snap point is used.
+    scene.core.mouse.setPosFromScenePoint(snapPoint);
+  }
+
   getSnapPoint(scene) {
     let snapPoint;
     let delta = 25 / scene.core.canvas.getScale(); // find a more suitable starting value
